Refactor the /tests as /e2etest
Moved the /tests to /e2etest directory, to have a better
distinguish between e2etests and unit tests that will be
added in the future.
