+++ .gitignore
@@ -1,1 +1,1 @@
-tests/artifacts/
+e2etest/artifacts/
