+++ e2etest/customEndpointSlices.json
